AG-8132 - Update dragAndDropService to account for scroll position
There are certain situations where the scroll position is not accounted for in mouse move event `pageX`/`pageY` values.

This was encountered when triggering synthetic `MouseMove` events on the homepage, where Chrome would not include the scroll position in the `pageX`/`pageY` calculation, but Firefox would. The update to `dragAndDropService` is made generic, so as to anticipate any other future situations where this unexpected edge case occurs.